+++ src/pages/LadderList.vue
@@ -1,0 +1,13 @@
+<template>
+    
+</template>
+
+<script>
+  export default {
+    name: "LadderList"
+  }
+</script>
+
+<style scoped lang="sass">
+    @import "../styles/ladderList"
+</style>
