+++ src/router/index.js
@@ -1,6 +1,8 @@
 import Vue from 'vue'
 import Router from 'vue-router'
 import App from '../App'
+import LadderDetail from '../pages/LadderDetail'
+import LadderList from '../pages/LadderList'
 
 Vue.use(Router)
 
@@ -8,8 +10,13 @@ export default new Router({
   routes: [
     {
       path: '/',
-      name: 'App',
-      component: App
+      name: 'list',
+      component: LadderList
+    },
+    {
+      path: '/detail',
+      name: 'detail',
+      component: LadderDetail
     }
   ]
 })
