+++ src/styles/_ladderDetail.sass
@@ -1,7 +1,3 @@
-.g-container
-  padding-top: 0 !important
-  padding-bottom: 0 !important
-  overflow: hidden
 .project-wrap
   position: relative
   top: 100px
