+++ src/styles/_layout.sass
@@ -1,0 +1,6 @@
+.g-container
+  overflow: hidden
+  padding-top: 0 !important
+  padding-bottom: 0 !important
+  display: flex
+  flex-direction: column
